drag_data_source/requirements.txt
Flask==2.3.3
numpy==2.2.6
PyYAML==6.0.1
tqdm==4.67.1
sentence-transformers>=2.3.0
rank-bm25>=0.2.2
faiss-cpu>=1.7.4
# For GPU support (uncomment if needed):
# faiss-gpu>=1.7.4
torch>=2.0.0
requests==2.32.5
jsonlines==4.0.0
# Blockchain/Web3 dependencies for contract interaction
web3>=6.0.0
eth-account>=0.8.0
